Split Governor into sub-modules
