+++ ouroboros-network/src/Ouroboros/Network/PeerSelection/Governor.hs
@@ -26,18 +26,13 @@ module Ouroboros.Network.PeerSelection.Governor (
     -- $peer-churn-governor
     peerChurnGovernor,
 
-    sanePeerSelectionTargets, --TODO: perhaps better to move to Types module
-    nullPeerSelectionTargets,
+    -- * Internals exported for testing
+    sanePeerSelectionTargets,
     PeerSelectionState(..),
 ) where
 
 import           Data.Void (Void)
-import           Data.Maybe (fromMaybe)
 import           Data.Semigroup (Min(..))
-import qualified Data.Map.Strict as Map
-import           Data.Map.Strict (Map)
-import qualified Data.Set as Set
-import           Data.Set (Set)
 
 import           Control.Applicative (Alternative((<|>)))
 import           Control.Monad.Class.MonadAsync
@@ -46,15 +41,17 @@ import           Control.Monad.Class.MonadSTM
 import           Control.Monad.Class.MonadTime
 import           Control.Monad.Class.MonadTimer
 import           Control.Tracer (Tracer(..), traceWith)
-import           Control.Exception (Exception(..), assert, SomeException)
-import           GHC.Stack
+import           Control.Exception (assert)
 
-import           Ouroboros.Network.PeerSelection.Types
 import qualified Ouroboros.Network.PeerSelection.KnownPeers as KnownPeers
-import           Ouroboros.Network.PeerSelection.KnownPeers (KnownPeers, KnownPeerInfo(..))
 import qualified Ouroboros.Network.PeerSelection.JobPool    as JobPool
-import           Ouroboros.Network.PeerSelection.JobPool (JobPool, Job(..))
-
+import           Ouroboros.Network.PeerSelection.JobPool (JobPool)
+import           Ouroboros.Network.PeerSelection.Governor.ActivePeers
+import           Ouroboros.Network.PeerSelection.Governor.EstablishedPeers
+import           Ouroboros.Network.PeerSelection.Governor.KnownPeers
+import           Ouroboros.Network.PeerSelection.Governor.Monitor
+import           Ouroboros.Network.PeerSelection.Governor.RootPeers
+import           Ouroboros.Network.PeerSelection.Governor.Types
 
 
 {- $overview
@@ -425,284 +422,6 @@ base our decision on include:
 -}
 
 
-data PeerSelectionPolicy peeraddr m = PeerSelectionPolicy {
-
-       policyPickKnownPeersForGossip :: PickPolicy peeraddr m,
-       policyPickColdPeersToPromote  :: PickPolicy peeraddr m,
-       policyPickWarmPeersToPromote  :: PickPolicy peeraddr m,
-       policyPickHotPeersToDemote    :: PickPolicy peeraddr m,
-       policyPickWarmPeersToDemote   :: PickPolicy peeraddr m,
-       policyPickColdPeersToForget   :: PickPolicy peeraddr m,
-
-       policyFindPublicRootTimeout   :: !DiffTime,
-       policyMaxInProgressGossipReqs :: !Int,
-       policyGossipRetryTime         :: !DiffTime,
-       policyGossipBatchWaitTime     :: !DiffTime,
-       policyGossipOverallTimeout    :: !DiffTime
-     }
-
--- | A peer pick policy is an action that picks a subset of elements from a
--- map of peers.
---
--- The pre-condition is that the map of available choices will be non-empty,
--- and the requested number to pick will be strictly positive.
---
--- The post-condition is that the picked set is non-empty but must not be
--- bigger than the requested number.
---
-type PickPolicy peeraddr m = Map peeraddr KnownPeerInfo
-                          -> Int
-                          -> STM m (Set peeraddr)
-
--- | Check pre-conditions and post-conditions on the pick policies
-pickPeers :: (Ord peeraddr, Functor m, HasCallStack)
-          => (Map peeraddr a -> Int -> m (Set peeraddr))
-          ->  Map peeraddr a -> Int -> m (Set peeraddr)
-pickPeers pick available num =
-    assert precondition $
-    fmap (\picked -> assert (postcondition picked) picked)
-         (pick available numClamped)
-  where
-    precondition         = not (Map.null available) && num > 0
-    postcondition picked = not (Set.null picked)
-                        && Set.size picked <= numClamped
-                        && picked `Set.isSubsetOf` Map.keysSet available
-    numClamped           = min num (Map.size available)
-
--- | Adjustable targets for the peer selection mechanism.
---
--- These are used by the peer selection governor as targets. They are used by
--- the peer churn governor loop as knobs to adjust, to influence the peer
--- selection governor.
---
--- The /known/, /established/ and /active/ peer targets are targets both from
--- below and from above: the governor will attempt to grow or shrink the sets
--- to hit these targets.
---
--- Unlike the other targets, the /root/ peer target is \"one sided\", it is
--- only a target from below. The governor does not try to shrink the root set
--- to hit it, it simply stops looking for more.
---
-data PeerSelectionTargets = PeerSelectionTargets {
-
-       targetNumberOfRootPeers        :: !Int,
-       targetNumberOfKnownPeers       :: !Int,
-       targetNumberOfEstablishedPeers :: !Int,
-       targetNumberOfActivePeers      :: !Int
-
-       -- Expressed as intervals rather than frequencies
---     targetChurnIntervalKnownPeers       :: !DiffTime,
---     targetChurnIntervalEstablishedPeers :: !DiffTime,
---     targetChurnIntervalActivePeers      :: !DiffTime
-     }
-  deriving (Eq, Show)
-
-nullPeerSelectionTargets :: PeerSelectionTargets
-nullPeerSelectionTargets =
-    PeerSelectionTargets {
-       targetNumberOfRootPeers        = 0,
-       targetNumberOfKnownPeers       = 0,
-       targetNumberOfEstablishedPeers = 0,
-       targetNumberOfActivePeers      = 0
---     targetChurnIntervalKnownPeers       = 0,
---     targetChurnIntervalEstablishedPeers = 0,
---     targetChurnIntervalActivePeers      = 0
-    }
-
-sanePeerSelectionTargets :: PeerSelectionTargets -> Bool
-sanePeerSelectionTargets PeerSelectionTargets{..} =
-                                 0 <= targetNumberOfActivePeers
- && targetNumberOfActivePeers      <= targetNumberOfEstablishedPeers
- && targetNumberOfEstablishedPeers <= targetNumberOfKnownPeers
- &&      targetNumberOfRootPeers   <= targetNumberOfKnownPeers
- && 0 <= targetNumberOfRootPeers
-
- && targetNumberOfActivePeers      <= 100
- && targetNumberOfEstablishedPeers <= 1000
- && targetNumberOfKnownPeers       <= 10000
-
-
--- | Actions performed by the peer selection governor.
---
--- These being pluggable allows:
---
--- * choice of known peer root sets
--- * running both in simulation and for real
---
-data PeerSelectionActions peeraddr peerconn m = PeerSelectionActions {
-
-       readPeerSelectionTargets :: STM m PeerSelectionTargets,
-
-       -- | Read the current set of locally or privately known root peers.
-       --
-       -- In general this is expected to be updated asynchronously by some
-       -- other thread. It is intended to cover the use case of peers from
-       -- local configuration. It could be dynamic due to DNS resolution, or
-       -- due to dynamic configuration updates.
-       --
-       readLocalRootPeers :: STM m (Map peeraddr PeerAdvertise),
-
-       -- | Request a sample of public root peers.
-       --
-       -- It is intended to cover use cases including:
-       --
-       -- * federated relays from a DNS pool
-       -- * stake pool relays published in the blockchain
-       -- * a pre-distributed snapshot of stake pool relays from the blockchain
-       --
-       requestPublicRootPeers :: Int -> m (Set peeraddr, DiffTime),
-
-       -- | The action to contact a known peer and request a sample of its
-       -- known peers.
-       --
-       -- This is synchronous, but it should expect to be interrupted by a
-       -- timeout asynchronous exception. Failures are throw as exceptions.
-       --
-       requestPeerGossip :: peeraddr -> m [peeraddr],
-
-       establishPeerConnection  :: peeraddr -> m peerconn,
-       monitorPeerConnection    :: peerconn -> STM m PeerStatus,
-       activatePeerConnection   :: peerconn -> m (),
-       deactivatePeerConnection :: peerconn -> m (),
-       closePeerConnection      :: peerconn -> m ()
-     }
-
-
--- | The internal state used by the 'peerSelectionGovernor'.
---
--- The local and public root sets are disjoint, and their union is the
--- overall root set.
---
-data PeerSelectionState peeraddr peerconn = PeerSelectionState {
-
-       targets              :: !PeerSelectionTargets,
-
-       -- | The current set of local root peers.
-       --
-       localRootPeers       :: !(Map peeraddr PeerAdvertise),
-
-       publicRootPeers      :: !(Set peeraddr),
-
-       -- |
-       --
-       knownPeers           :: !(KnownPeers peeraddr),
-
-       -- |
-       --
-       establishedPeers     :: !(Map peeraddr peerconn),
-       establishedStatus    :: !(Map peeraddr PeerStatus),
-
-       -- |
-       --
-       activePeers          :: !(Set peeraddr),
-
-       -- | A counter to manage the exponential backoff strategy for when to
-       -- retry querying for more public root peers. It is negative for retry
-       -- counts after failure, and positive for retry counts that are
-       -- successful but make no progress.
-       --
-       publicRootBackoffs   :: !Int,
-
-       -- | The earliest time we would be prepared to request more public root
-       -- peers. This is used with the 'publicRootBackoffs' to manage the
-       -- exponential backoff.
-       --
-       publicRootRetryTime  :: !Time,
-
-       inProgressPublicRootsReq :: !Bool,
-       inProgressGossipReqs     :: !Int,
-       inProgressPromoteCold    :: !(Set peeraddr),
-       inProgressPromoteWarm    :: !(Set peeraddr),
-       inProgressDemoteWarm     :: !(Set peeraddr),
-       inProgressDemoteHot      :: !(Set peeraddr)
-
---     TODO: need something like this to distinguish between lots of bad peers
---     and us getting disconnected from the network locally. We don't want a
---     network disconnect to cause us to flush our full known peer set by
---     considering them all to have bad connectivity.
---     Should also take account of DNS failures for root peer set.
---     lastSucessfulNetworkEvent :: Time
-     }
-  deriving (Show, Functor)
-
-emptyPeerSelectionState :: PeerSelectionState peeraddr peerconn
-emptyPeerSelectionState =
-    PeerSelectionState {
-      targets              = nullPeerSelectionTargets,
-      localRootPeers       = Map.empty,
-      publicRootPeers      = Set.empty,
-      knownPeers           = KnownPeers.empty,
-      establishedPeers     = Map.empty,
-      establishedStatus    = Map.empty,
-      activePeers          = Set.empty,
-      publicRootBackoffs   = 0,
-      publicRootRetryTime  = Time 0,
-      inProgressPublicRootsReq = False,
-      inProgressGossipReqs     = 0,
-      inProgressPromoteCold    = Set.empty,
-      inProgressPromoteWarm    = Set.empty,
-      inProgressDemoteWarm     = Set.empty,
-      inProgressDemoteHot      = Set.empty
-    }
-
-invariantPeerSelectionState :: Ord peeraddr
-                            => PeerSelectionState peeraddr peerconn -> Bool
-invariantPeerSelectionState PeerSelectionState{..} =
-    KnownPeers.invariant knownPeers
-
-    -- The activePeers is a subset of the establishedPeers
-    -- which is a subset of the known peers
- && Set.isSubsetOf activePeersSet establishedPeersSet
- && Set.isSubsetOf establishedPeersSet knownPeersSet
- && Map.keysSet establishedStatus == establishedPeersSet
-
-    -- The localRootPeers and publicRootPeers must not overlap.
- && Set.null (Set.intersection localRootPeersSet publicRootPeers)
-
-    -- The localRootPeers are a subset of the knownPeers,
-    -- and with correct source and other info in the knownPeers.
- && Map.isSubmapOfBy (\rootPeerAdvertise
-                       KnownPeerInfo {knownPeerAdvertise, knownPeerSource} ->
-                           knownPeerSource == PeerSourceLocalRoot
-                        && knownPeerAdvertise == rootPeerAdvertise)
-                     localRootPeers
-                     (KnownPeers.toMap knownPeers)
-
-    -- The publicRootPeers are a subset of the knownPeers,
-    -- and with correct source info in the knownPeers.
- && Map.isSubmapOfBy (\_ KnownPeerInfo {knownPeerSource} ->
-                         knownPeerSource == PeerSourcePublicRoot)
-                     (Map.fromSet (const ()) publicRootPeers)
-                     (KnownPeers.toMap knownPeers)
-
-    --TODO: all other peers have PeerSourceGossip, so no stale source info.
-
-    -- We don't want to pick local root peers to forget, so it had better be
-    -- the case that there's fewer of them than our target number.
- && Map.size localRootPeers <= targetNumberOfKnownPeers targets
-
-    -- All currently established peers are in the availableToConnect set since
-    -- the alternative is a record of failure, but these are not (yet) failed.
- && Set.isSubsetOf establishedPeersSet (KnownPeers.availableToConnect knownPeers)
-
-    -- No constraint for publicRootBackoffs, publicRootRetryTime
-    -- or inProgressPublicRootsReq
-
- && inProgressGossipReqs >= 0
- && Set.isSubsetOf inProgressPromoteCold coldPeersSet
- && Set.isSubsetOf inProgressPromoteWarm warmPeersSet
- && Set.isSubsetOf inProgressDemoteWarm  warmPeersSet
- && Set.isSubsetOf inProgressDemoteHot   hotPeersSet
- && Set.null (Set.intersection inProgressPromoteWarm inProgressDemoteWarm)
-  where
-    localRootPeersSet   = Map.keysSet localRootPeers
-    knownPeersSet       = Map.keysSet (KnownPeers.toMap knownPeers)
-    establishedPeersSet = Map.keysSet establishedPeers
-    activePeersSet      = activePeers
-    coldPeersSet        = knownPeersSet Set.\\ establishedPeersSet
-    warmPeersSet        = establishedPeersSet Set.\\ activePeersSet
-    hotPeersSet         = activePeersSet
-
 -- |
 --
 peerSelectionGovernor :: (MonadAsync m, MonadMask m, MonadTime m, MonadTimer m,
@@ -818,30 +537,6 @@ peerSelectionGovernorLoop tracer debugTracer
       -- roots peers because we'd be above target for known peers).
 
 
-data Guarded m a = GuardedSkip !(Maybe (Min Time))
-                 | Guarded     !(Maybe (Min Time)) (m a)
-
-instance Alternative m => Semigroup (Guarded m a) where
-  Guarded     ta a <> Guarded     tb b = Guarded     (ta <> tb) (a <|> b)
-  Guarded     ta a <> GuardedSkip tb   = Guarded     (ta <> tb)  a
-  GuardedSkip ta   <> Guarded     tb b = Guarded     (ta <> tb)  b
-  GuardedSkip ta   <> GuardedSkip tb   = GuardedSkip (ta <> tb)
-
-
-data Decision m peeraddr peerconn = Decision {
-         -- | A trace event to classify the decision and action
-       decisionTrace :: TracePeerSelection peeraddr,
-
-         -- | An updated state to use immediately
-       decisionState :: PeerSelectionState peeraddr peerconn,
-
-       -- | An optional 'Job' to execute asynchronously. This job leads to
-       -- a further 'Decision'. This gives a state update to apply upon
-       -- completion, but also allows chaining further job actions.
-       --
-       decisionJobs  :: [Job m (Completion m peeraddr peerconn)]
-     }
-
 wakeupDecision :: PeerSelectionState peeraddr peerconn
                -> Decision m peeraddr peerconn
 wakeupDecision st =
@@ -850,1058 +545,6 @@ wakeupDecision st =
     decisionState = st,
     decisionJobs  = []
   }
-
-newtype Completion m peeraddr peerconn =
-        Completion (PeerSelectionState peeraddr peerconn
-                 -> Time -> Decision m peeraddr peerconn)
-
-data TracePeerSelection peeraddr =
-       TraceLocalRootPeersChanged (Map peeraddr PeerAdvertise)
-                                  (Map peeraddr PeerAdvertise)
-     | TraceTargetsChanged     PeerSelectionTargets PeerSelectionTargets
-     | TracePublicRootsRequest Int Int
-     | TracePublicRootsResults (Set peeraddr) Int DiffTime
-     | TracePublicRootsFailure SomeException Int DiffTime
-     | TraceGossipRequests     Int Int (Set peeraddr) (Set peeraddr) -- target, actual, selected
-     | TraceGossipResults      [(peeraddr, Either SomeException [peeraddr])] --TODO: classify failures
-     | TraceForgetColdPeers    Int Int (Set peeraddr) -- target, actual, selected
-     | TracePromoteColdPeers   Int Int (Set peeraddr)
-     | TracePromoteColdFailed  peeraddr SomeException
-     | TracePromoteColdDone    peeraddr
-     | TracePromoteWarmPeers   Int Int (Set peeraddr)
-     | TracePromoteWarmFailed  peeraddr SomeException
-     | TracePromoteWarmDone    peeraddr
-     | TraceDemoteWarmPeers    Int Int (Set peeraddr) -- target, actual, selected
-     | TraceDemoteWarmFailed   peeraddr SomeException
-     | TraceDemoteWarmDone     peeraddr
-     | TraceDemoteHotPeers     Int Int (Set peeraddr)
-     | TraceDemoteHotFailed    peeraddr SomeException
-     | TraceDemoteHotDone      peeraddr
-     | TraceDemoteAsynchronous (Map peeraddr PeerStatus)
-     | TraceGovernorWakeup
-  deriving Show
-
-data DebugPeerSelection peeraddr peerconn =
-       TraceGovernorState  (PeerSelectionState peeraddr peerconn)
-                           (Maybe DiffTime)
-  deriving (Show, Functor)
-
-rootPeersBelowTarget :: (MonadSTM m, Ord peeraddr)
-                     => PeerSelectionActions peeraddr peerconn m
-                     -> PeerSelectionState peeraddr peerconn
-                     -> Time
-                     -> Guarded (STM m) (Decision m peeraddr peerconn)
-rootPeersBelowTarget actions
-                     st@PeerSelectionState {
-                       localRootPeers,
-                       publicRootPeers,
-                       publicRootRetryTime,
-                       inProgressPublicRootsReq,
-                       targets = PeerSelectionTargets {
-                                   targetNumberOfRootPeers
-                                 }
-                     }
-                     now
-    -- Are we under target for number of root peers?
-  | maxExtraRootPeers > 0
-
-    -- Are we already requesting more root peers?
-  , not inProgressPublicRootsReq
-
-    -- We limit how frequently we make requests, are we allowed to do it yet?
-  , now >= publicRootRetryTime
-  = Guarded Nothing $
-      return Decision {
-        decisionTrace = TracePublicRootsRequest
-                          targetNumberOfRootPeers
-                          numRootPeers,
-        decisionState = st { inProgressPublicRootsReq = True },
-        decisionJobs  = [jobReqPublicRootPeers actions maxExtraRootPeers]
-      }
-
-    -- If we would be able to do the request except for the time, return the
-    -- next retry time.
-  | maxExtraRootPeers > 0
-  , not inProgressPublicRootsReq
-  = GuardedSkip (Just (Min publicRootRetryTime))
-
-  | otherwise
-  = GuardedSkip Nothing
-  where
-    numRootPeers      = Map.size localRootPeers + Set.size publicRootPeers
-    maxExtraRootPeers = targetNumberOfRootPeers - numRootPeers
-
-jobReqPublicRootPeers :: forall m peeraddr peerconn.
-                         (Monad m, Ord peeraddr)
-                      => PeerSelectionActions peeraddr peerconn m
-                      -> Int
-                      -> Job m (Completion m peeraddr peerconn)
-jobReqPublicRootPeers PeerSelectionActions{requestPublicRootPeers}
-                   numExtraAllowed =
-    Job job handler
-  where
-    handler :: SomeException -> Completion m peeraddr peerconn
-    handler e =
-      Completion $ \st now ->
-      -- This is a failure, so move the backoff counter one in the failure
-      -- direction (negative) and schedule the next retry time accordingly.
-      -- We use an exponential backoff strategy. The max retry time of 2^12
-      -- seconds is just over an hour.
-      let publicRootBackoffs'      :: Int
-          publicRootBackoffs'      = (publicRootBackoffs st `min` 0) - 1
-
-          publicRootRetryDiffTime' :: DiffTime
-          publicRootRetryDiffTime' = 2 ^ (abs publicRootBackoffs' `min` 12)
-
-          publicRootRetryTime'     :: Time
-          publicRootRetryTime'     = addTime publicRootRetryDiffTime' now
-       in Decision {
-            decisionTrace = TracePublicRootsFailure
-                              e
-                              publicRootBackoffs'
-                              publicRootRetryDiffTime',
-            decisionState = st {
-                              inProgressPublicRootsReq = False,
-                              publicRootBackoffs  = publicRootBackoffs',
-                              publicRootRetryTime = publicRootRetryTime'
-                            },
-            decisionJobs  = []
-          }
-
-    job :: m (Completion m peeraddr peerconn)
-    job = do
-      (results, ttl) <- requestPublicRootPeers numExtraAllowed
-      return $ Completion $ \st now ->
-        let newPeers         = results Set.\\ Map.keysSet (localRootPeers st)
-                                       Set.\\ publicRootPeers st
-            publicRootPeers' = publicRootPeers st <> newPeers
-            knownPeers'      = KnownPeers.insert
-                                 PeerSourcePublicRoot
-                                 (const DoAdvertisePeer)
-                                 newPeers
-                                 (knownPeers st)
-
-            -- We got a successful response to our request, but if we're still
-            -- below target we're going to want to try again at some point.
-            -- If we made progress towards our target then we will retry at the
-            -- suggested ttl. But if we did not make progress then we want to
-            -- follow an exponential backoff strategy. The max retry time of 2^12
-            -- seconds is just over an hour.
-            publicRootBackoffs' :: Int
-            publicRootBackoffs'
-              | Set.null newPeers = (publicRootBackoffs st `max` 0) + 1
-              | otherwise         = 0
-
-            publicRootRetryDiffTime :: DiffTime
-            publicRootRetryDiffTime
-              | publicRootBackoffs' == 0
-                          = ttl
-              | otherwise = 2^(publicRootBackoffs' `min` 12)
-
-            publicRootRetryTime :: Time
-            publicRootRetryTime = addTime publicRootRetryDiffTime now
-         in Decision {
-              decisionTrace = TracePublicRootsResults
-                                newPeers
-                                publicRootBackoffs'
-                                publicRootRetryDiffTime,
-              decisionState = st {
-                                publicRootPeers     = publicRootPeers',
-                                knownPeers          = knownPeers',
-                                publicRootBackoffs  = publicRootBackoffs',
-                                publicRootRetryTime = publicRootRetryTime,
-                                inProgressPublicRootsReq = False
-                              },
-              decisionJobs  = []
-            }
-
-
-knownPeersBelowTarget :: (MonadAsync m, MonadTimer m, Ord peeraddr)
-                      => PeerSelectionActions peeraddr peerconn m
-                      -> PeerSelectionPolicy peeraddr m
-                      -> PeerSelectionState peeraddr peerconn
-                      -> Time
-                      -> Guarded (STM m) (Decision m peeraddr peerconn)
-knownPeersBelowTarget actions
-                      policy@PeerSelectionPolicy {
-                        policyMaxInProgressGossipReqs,
-                        policyPickKnownPeersForGossip,
-                        policyGossipRetryTime
-                      }
-                      st@PeerSelectionState {
-                        knownPeers,
-                        inProgressGossipReqs,
-                        targets = PeerSelectionTargets {
-                                    targetNumberOfKnownPeers
-                                  }
-                      }
-                      now
-    -- Are we under target for number of known peers?
-  | numKnownPeers < targetNumberOfKnownPeers
-
-    -- Are we at our limit for number of gossip requests?
-  , numGossipReqsPossible > 0
-
-    -- Are there any known peers that we can send a gossip request to?
-    -- We can only ask ones where we have not asked them within a certain time.
-  , not (Set.null availableForGossip)
-  = Guarded Nothing $ do
-      selectedForGossip <- pickPeers
-                             policyPickKnownPeersForGossip
-                             (KnownPeers.toMap knownPeers
-                                `Map.restrictKeys` availableForGossip)
-                             numGossipReqsPossible
-      let numGossipReqs = Set.size selectedForGossip
-      return Decision {
-        decisionTrace = TraceGossipRequests
-                          targetNumberOfKnownPeers
-                          numKnownPeers
-                          availableForGossip
-                          selectedForGossip,
-        decisionState = st {
-                          inProgressGossipReqs = inProgressGossipReqs
-                                               + numGossipReqs,
-                          knownPeers = KnownPeers.setGossipTime
-                                         selectedForGossip
-                                         (addTime policyGossipRetryTime now)
-                                         knownPeers
-                        },
-        decisionJobs  = [jobGossip actions policy
-                           (Set.toList selectedForGossip)]
-      }
-
-    -- If we could gossip except that there are none currently available
-    -- then we return the next wakeup time (if any)
-  | numKnownPeers < targetNumberOfKnownPeers
-  , numGossipReqsPossible > 0
-  , Set.null availableForGossip
-  = GuardedSkip (Min <$> KnownPeers.minGossipTime knownPeers)
-
-  | otherwise
-  = GuardedSkip Nothing
-  where
-    numKnownPeers         = KnownPeers.size knownPeers
-    numGossipReqsPossible = policyMaxInProgressGossipReqs
-                          - inProgressGossipReqs
-    availableForGossip    = KnownPeers.availableForGossip knownPeers
-
-
-jobGossip :: forall m peeraddr peerconn.
-             (MonadAsync m, MonadTimer m, Ord peeraddr)
-          => PeerSelectionActions peeraddr peerconn m
-          -> PeerSelectionPolicy peeraddr m
-          -> [peeraddr]
-          -> Job m (Completion m peeraddr peerconn)
-jobGossip PeerSelectionActions{requestPeerGossip}
-           PeerSelectionPolicy{..} =
-    \peers -> Job (jobPhase1 peers) (handler peers)
-  where
-    handler :: [peeraddr] -> SomeException -> Completion m peeraddr peerconn
-    handler peers e =
-      Completion $ \st _ ->
-      Decision {
-        decisionTrace = TraceGossipResults [ (p, Left e) | p <- peers ],
-        decisionState = st {
-                          inProgressGossipReqs = inProgressGossipReqs st
-                                               - length peers
-                        },
-        decisionJobs  = []
-      }
-
-    jobPhase1 :: [peeraddr] -> m (Completion m peeraddr peerconn)
-    jobPhase1 peers = do
-      -- In the typical case, where most requests return within a short
-      -- timeout we want to collect all the responses into a batch and
-      -- add them to the known peers set in one go.
-      --
-      -- So fire them all off in one go:
-      gossips <- sequence [ async (requestPeerGossip peer) | peer <- peers ]
-
-      -- First to finish synchronisation between /all/ the gossips completing
-      -- or the timeout (with whatever partial results we have at the time)
-      results <- waitAllCatchOrTimeout gossips policyGossipBatchWaitTime
-      case results of
-        Right totalResults -> do
-          let peerResults = zip peers totalResults
-              newPeers    = [ p | Right ps <- totalResults, p <- ps ]
-          return $ Completion $ \st _ -> Decision {
-            decisionTrace = TraceGossipResults peerResults,
-            decisionState = st {
-                              --TODO: also update with the failures
-                              knownPeers = KnownPeers.insert
-                                             PeerSourceGossip
-                                             (const DoAdvertisePeer)
-                                             (Set.fromList newPeers)
-                                             (knownPeers st),
-                              inProgressGossipReqs = inProgressGossipReqs st
-                                                   - length peers
-                            },
-            decisionJobs  = []
-          }
-
-        -- But if any don't make the first timeout then they'll be added later
-        -- when they do reply or never if we hit the hard timeout.
-        Left partialResults -> do
-
-          -- We have to keep track of the relationship between the peer
-          -- addresses and the gossip requests, completed and still in progress:
-          let peerResults      = [ (p, r)
-                                 | (p, Just r)  <- zip peers   partialResults ]
-              newPeers         = [  p
-                                 | Just (Right ps) <-          partialResults
-                                 ,  p <- ps ]
-              peersRemaining   = [  p
-                                 | (p, Nothing) <- zip peers   partialResults ]
-              gossipsRemaining = [  a
-                                 | (a, Nothing) <- zip gossips partialResults ]
-
-          return $ Completion $ \st _ -> Decision {
-            decisionTrace = TraceGossipResults peerResults,
-            decisionState = st {
-                              --TODO: also update with the failures
-                              knownPeers = KnownPeers.insert
-                                             PeerSourceGossip
-                                             (const DoAdvertisePeer)
-                                             (Set.fromList newPeers)
-                                             (knownPeers st),
-                              inProgressGossipReqs = inProgressGossipReqs st
-                                                   - length peerResults
-                            },
-            decisionJobs  = [Job (jobPhase2 peersRemaining gossipsRemaining)
-                                 (handler peersRemaining)]
-          }
-
-    jobPhase2 :: [peeraddr] -> [Async m [peeraddr]]
-              -> m (Completion m peeraddr peerconn)
-    jobPhase2 peers gossips = do
-
-      -- Wait again, for all remaining to finish or a timeout.
-      results <- waitAllCatchOrTimeout
-                      gossips
-                      (policyGossipOverallTimeout
-                       - policyGossipBatchWaitTime)
-      let peerResults =
-            case results of
-              Right totalResults  -> zip peers totalResults
-              Left partialResults -> [ (p, fromMaybe err r)
-                                     | (p, r) <- zip peers partialResults ]
-                where err = Left (toException AsyncCancelled)
-
-          newPeers =
-            case results of
-              Right totalResults  -> [ p | Right ps <- totalResults,  p <- ps ]
-              Left partialResults -> [ p | Just (Right ps) <- partialResults,  p <- ps ]
-
-          gossipsIncomplete =
-            case results of
-              Right _totalResults -> []
-              Left partialResults ->
-                [ a | (a, Nothing) <- zip gossips partialResults ]
-
-      mapM_ cancel gossipsIncomplete
-
-      return $ Completion $ \st _ -> Decision {
-        decisionTrace = TraceGossipResults peerResults,
-        decisionState = st {
-                          --TODO: also update with the failures
-                          knownPeers = KnownPeers.insert
-                                         PeerSourceGossip
-                                         (const DoAdvertisePeer)
-                                         (Set.fromList newPeers)
-                                         (knownPeers st),
-                          inProgressGossipReqs = inProgressGossipReqs st
-                                               - length peers
-                        },
-        decisionJobs  = []
-      }
-
-
-knownPeersAboveTarget :: (MonadSTM m, Ord peeraddr)
-                      => PeerSelectionPolicy peeraddr m
-                      -> PeerSelectionState peeraddr peerconn
-                      -> Guarded (STM m) (Decision m peeraddr peerconn)
-knownPeersAboveTarget PeerSelectionPolicy {
-                        policyPickColdPeersToForget
-                      }
-                      st@PeerSelectionState {
-                        localRootPeers,
-                        publicRootPeers,
-                        knownPeers,
-                        establishedPeers,
-                        inProgressPromoteCold,
-                        targets = PeerSelectionTargets {
-                                    targetNumberOfKnownPeers,
-                                    targetNumberOfRootPeers
-                                  }
-                      }
-    -- Are we above the target for number of known peers?
-  | numKnownPeers > targetNumberOfKnownPeers
-
-    -- Are there any cold peers we could pick to forget?
-    -- As a first cheap approximation, check if there are any cold peers.
-  , numKnownPeers > numEstablishedPeers
-
-    -- Beyond this it gets more complicated, and it is not clear that there
-    -- are any precise cheap checks. So we just do the full calculation.
-    -- In particular there can be overlap between cold peers and root peers
-    -- and we have constraints on forgetting root peers.
-    --
-    -- We must never pick local root peers to forget as this would violate
-    -- our invariant that the localRootPeers is a subset of the knownPeers.
-    --
-    -- We also need to avoid picking public root peers if that would put us
-    -- below the target for root peers.
-    --
-  , let numRootPeersCanForget = Map.size localRootPeers
-                              + Set.size publicRootPeers
-                              - targetNumberOfRootPeers
-        protectedRootPeers    = Map.keysSet localRootPeers
-                             <> Set.drop numRootPeersCanForget publicRootPeers
-        availableToForget     = KnownPeers.toMap knownPeers
-                                   Map.\\ establishedPeers
-                                  `Map.withoutKeys` protectedRootPeers
-                                  `Map.withoutKeys` inProgressPromoteCold
-
-  , not (Map.null availableToForget)
-  = Guarded Nothing $ do
-      let numPeersToForget = numKnownPeers - targetNumberOfKnownPeers
-      selectedToForget <- pickPeers
-                            policyPickColdPeersToForget
-                            availableToForget
-                            numPeersToForget
-      return Decision {
-        decisionTrace = TraceForgetColdPeers
-                          targetNumberOfKnownPeers
-                          numKnownPeers
-                          selectedToForget,
-        decisionState = st {
-                          knownPeers      = KnownPeers.delete
-                                              selectedToForget
-                                              knownPeers,
-                          publicRootPeers = publicRootPeers
-                                              Set.\\ selectedToForget
-                        },
-        decisionJobs  = []
-      }
-
-  | otherwise
-  = GuardedSkip Nothing
-  where
-    numKnownPeers, numEstablishedPeers :: Int
-    numKnownPeers        = KnownPeers.size knownPeers
-    numEstablishedPeers  = Map.size establishedPeers
-
-
-establishedPeersBelowTarget :: forall peeraddr peerconn m.
-                               (MonadSTM m, Ord peeraddr)
-                            => PeerSelectionActions peeraddr peerconn m
-                            -> PeerSelectionPolicy peeraddr m
-                            -> PeerSelectionState peeraddr peerconn
-                            -> Guarded (STM m) (Decision m peeraddr peerconn)
-establishedPeersBelowTarget actions
-                            PeerSelectionPolicy {
-                              policyPickColdPeersToPromote
-                            }
-                            st@PeerSelectionState {
-                              knownPeers,
-                              establishedPeers,
-                              inProgressPromoteCold,
-                              targets = PeerSelectionTargets {
-                                          targetNumberOfEstablishedPeers
-                                        }
-                            }
-    -- Are we below the target for number of established peers?
-  | numEstablishedPeers + numConnectInProgress < targetNumberOfEstablishedPeers
-
-    -- Are there any cold peers we could possibly pick to connect to?
-    -- We can subtract the established ones because by definition they are
-    -- not cold and our invariant is that they are always in the connect set.
-    -- We can also subtract the in progress ones since they are also already
-    -- in the connect set and we cannot pick them again.
-  , Set.size availableToConnect - numEstablishedPeers - numConnectInProgress > 0
-  = Guarded Nothing $ do
-      -- The availableToPromote here is non-empty due to the second guard.
-      -- The known peers map restricted to the connect set is the same size as
-      -- the connect set (because it is a subset). The establishedPeers is a
-      -- subset of the connect set and we also know that there is no overlap
-      -- between inProgressPromoteCold and establishedPeers. QED.
-      --
-      -- The numPeersToPromote is positive based on the first guard.
-      --
-      let availableToPromote :: Map peeraddr KnownPeerInfo
-          availableToPromote = KnownPeers.toMap knownPeers
-                                `Map.restrictKeys` availableToConnect
-                                 Map.\\ establishedPeers
-                                `Map.withoutKeys` inProgressPromoteCold
-          numPeersToPromote  = targetNumberOfEstablishedPeers
-                             - numEstablishedPeers
-                             - numConnectInProgress
-      selectedToPromote <- pickPeers
-                             policyPickColdPeersToPromote
-                             availableToPromote
-                             numPeersToPromote
-      return Decision {
-        decisionTrace = TracePromoteColdPeers
-                          targetNumberOfEstablishedPeers
-                          numEstablishedPeers
-                          selectedToPromote,
-        decisionState = st {
-                          inProgressPromoteCold = inProgressPromoteCold
-                                               <> selectedToPromote
-                        },
-        decisionJobs  = [ jobPromoteColdPeer actions peer
-                        | peer <- Set.toList selectedToPromote ]
-      }
-
-    -- If we could connect except that there are no peers currently available
-    -- then we return the next wakeup time (if any)
-  | numEstablishedPeers + numConnectInProgress < targetNumberOfEstablishedPeers
-  = GuardedSkip (Min <$> KnownPeers.minConnectTime knownPeers)
-
-  | otherwise
-  = GuardedSkip Nothing
-  where
-    numEstablishedPeers, numConnectInProgress :: Int
-    numEstablishedPeers  = Map.size establishedPeers
-    numConnectInProgress = Set.size inProgressPromoteCold
-    availableToConnect   = KnownPeers.availableToConnect knownPeers
-
-jobPromoteColdPeer :: forall peeraddr peerconn m.
-                       (Monad m, Ord peeraddr)
-                   => PeerSelectionActions peeraddr peerconn m
-                   -> peeraddr
-                   -> Job m (Completion m peeraddr peerconn)
-jobPromoteColdPeer PeerSelectionActions{establishPeerConnection} peeraddr =
-    Job job handler
-  where
-    handler :: SomeException -> Completion m peeraddr peerconn
-    handler e =
-      Completion $ \st _now -> Decision {
-        decisionTrace = TracePromoteColdFailed peeraddr e,
-        decisionState = st {
-                          knownPeers            = KnownPeers.incrementFailCount
-                                                    peeraddr (knownPeers st),
-                          inProgressPromoteCold = Set.delete peeraddr
-                                                    (inProgressPromoteCold st)
-                        },
-        decisionJobs  = []
-      }
-
-    job :: m (Completion m peeraddr peerconn)
-    job = do
-      --TODO: decide if we should do timeouts here or if we should make that
-      -- the responsibility of establishPeerConnection
-      peerconn <- establishPeerConnection peeraddr
-      return $ Completion $ \st _now -> Decision {
-        decisionTrace = TracePromoteColdDone peeraddr,
-        decisionState = st {
-                          establishedPeers      = Map.insert peeraddr peerconn
-                                                    (establishedPeers st),
-                          establishedStatus     = Map.insert peeraddr PeerWarm
-                                                    (establishedStatus st),
-                          inProgressPromoteCold = Set.delete peeraddr
-                                                    (inProgressPromoteCold st)
-                        },
-        decisionJobs  = []
-      }
-
-
-establishedPeersAboveTarget :: forall peeraddr peerconn m.
-                               (MonadSTM m, Ord peeraddr)
-                            => PeerSelectionActions peeraddr peerconn m
-                            -> PeerSelectionPolicy peeraddr m
-                            -> PeerSelectionState peeraddr peerconn
-                            -> Guarded (STM m) (Decision m peeraddr peerconn)
-establishedPeersAboveTarget actions
-                            PeerSelectionPolicy {
-                              policyPickWarmPeersToDemote
-                            }
-                            st@PeerSelectionState {
-                              knownPeers,
-                              establishedPeers,
-                              activePeers,
-                              inProgressDemoteWarm,
-                              inProgressPromoteWarm,
-                              targets = PeerSelectionTargets {
-                                          targetNumberOfEstablishedPeers
-                                        }
-                            }
-    -- Are we above the target for number of established peers?
-    -- Or more precisely, how many established peers could we demote?
-    -- We only want to pick established peers that are not active, since for
-    -- active one we need to demote them first.
-  | let numEstablishedPeers, numActivePeers, numPeersToDemote :: Int
-        numEstablishedPeers = Map.size establishedPeers
-        numActivePeers      = Set.size activePeers
-        -- One constraint on how many to demote is the difference in the
-        -- number we have now vs the target. The other constraint is that
-        -- we pick established peers that are not also active. These
-        -- constraints combine by taking the minimum. We must also subtract
-        -- the number we're demoting so we don't repeat the same work. And
-        -- cannot demote ones we're in the process of promoting.
-        numPeersToDemote    = min (numEstablishedPeers
-                                   - targetNumberOfEstablishedPeers)
-                                  (numEstablishedPeers
-                                   - numActivePeers)
-                            - Set.size inProgressDemoteWarm
-                            - Set.size inProgressPromoteWarm
-  , numPeersToDemote > 0
-  = Guarded Nothing $ do
-
-      let availableToDemote :: Map peeraddr KnownPeerInfo
-          availableToDemote = KnownPeers.toMap knownPeers
-                               `Map.intersection` establishedPeers
-                               `Map.withoutKeys` activePeers
-                               `Map.withoutKeys` inProgressDemoteWarm
-                               `Map.withoutKeys` inProgressPromoteWarm
-      selectedToDemote <- pickPeers
-                            policyPickWarmPeersToDemote
-                            availableToDemote
-                            numPeersToDemote
-      let selectedToDemote' :: Map peeraddr peerconn
-          selectedToDemote' = establishedPeers
-                                `Map.restrictKeys` selectedToDemote
-
-      return Decision {
-        decisionTrace = TraceDemoteWarmPeers
-                          targetNumberOfEstablishedPeers
-                          numEstablishedPeers
-                          selectedToDemote,
-        decisionState = st {
-                          inProgressDemoteWarm = inProgressDemoteWarm
-                                              <> selectedToDemote
-                        },
-        decisionJobs  = [ jobDemoteEstablishedPeer actions peeraddr peerconn
-                        | (peeraddr, peerconn) <- Map.assocs selectedToDemote' ]
-      }
-
-  | otherwise
-  = GuardedSkip Nothing
-
-jobDemoteEstablishedPeer :: forall peeraddr peerconn m.
-                            (Monad m, Ord peeraddr)
-                         => PeerSelectionActions peeraddr peerconn m
-                         -> peeraddr
-                         -> peerconn
-                         -> Job m (Completion m peeraddr peerconn)
-jobDemoteEstablishedPeer PeerSelectionActions{closePeerConnection}
-                         peeraddr peerconn =
-    Job job handler
-  where
-    handler :: SomeException -> Completion m peeraddr peerconn
-    handler e =
-      -- It's quite bad if closing fails, but the best we can do is revert to
-      -- the state where we believed this peer is still warm, since then we
-      -- can have another go or perhaps it'll be closed for other reasons and
-      -- our monitoring will notice it.
-      Completion $ \st _now -> Decision {
-        decisionTrace = TraceDemoteWarmFailed peeraddr e,
-        decisionState = st {
-                          inProgressDemoteWarm = Set.delete peeraddr
-                                                   (inProgressDemoteWarm st)
-                        },
-        decisionJobs  = []
-      }
-
-    job :: m (Completion m peeraddr peerconn)
-    job = do
-      closePeerConnection peerconn
-      return $ Completion $ \st _now -> Decision {
-        decisionTrace = TraceDemoteWarmDone peeraddr,
-        decisionState = st {
-                          establishedPeers     = Map.delete peeraddr
-                                                   (establishedPeers st),
-                          establishedStatus    = Map.delete peeraddr
-                                                   (establishedStatus st),
-                          inProgressDemoteWarm = Set.delete peeraddr
-                                                   (inProgressDemoteWarm st)
-                        },
-        decisionJobs  = []
-      }
-
-activePeersBelowTarget :: forall peeraddr peerconn m.
-                          (MonadSTM m, Ord peeraddr)
-                       => PeerSelectionActions peeraddr peerconn m
-                       -> PeerSelectionPolicy peeraddr m
-                       -> PeerSelectionState peeraddr peerconn
-                       -> Guarded (STM m) (Decision m peeraddr peerconn)
-activePeersBelowTarget actions
-                       PeerSelectionPolicy {
-                         policyPickWarmPeersToPromote
-                       }
-                       st@PeerSelectionState {
-                         knownPeers,
-                         establishedPeers,
-                         activePeers,
-                         inProgressPromoteWarm,
-                         inProgressDemoteWarm,
-                         targets = PeerSelectionTargets {
-                                     targetNumberOfActivePeers
-                                   }
-                       }
-    -- Are we below the target for number of active peers?
-  | numActivePeers + numPromoteInProgress < targetNumberOfActivePeers
-
-    -- Are there any warm peers we could pick to promote?
-  , numEstablishedPeers - numActivePeers
-                        - numPromoteInProgress - numDemoteInProgress > 0
-  = Guarded Nothing $ do
-          -- The availableToPromote is non-empty due to the second guard.
-          -- The numPeersToPromote is positive due to the first guard.
-      let availableToPromote :: Map peeraddr KnownPeerInfo
-          availableToPromote = KnownPeers.toMap knownPeers
-                                `Map.intersection` establishedPeers
-                                `Map.withoutKeys` activePeers
-                                `Map.withoutKeys` inProgressPromoteWarm
-                                `Map.withoutKeys` inProgressDemoteWarm
-          numPeersToPromote  = targetNumberOfActivePeers
-                             - numActivePeers
-                             - numPromoteInProgress
-      selectedToPromote <- pickPeers
-                             policyPickWarmPeersToPromote
-                             availableToPromote
-                             numPeersToPromote
-      let selectedToPromote' :: Map peeraddr peerconn
-          selectedToPromote' = establishedPeers
-                                 `Map.restrictKeys` selectedToPromote
-      return Decision {
-        decisionTrace = TracePromoteWarmPeers
-                          targetNumberOfActivePeers
-                          numActivePeers
-                          selectedToPromote,
-        decisionState = st {
-                          inProgressPromoteWarm = inProgressPromoteWarm
-                                               <> selectedToPromote
-                        },
-        decisionJobs  = [ jobPromoteWarmPeer actions peeraddr peerconn
-                        | (peeraddr, peerconn) <- Map.assocs selectedToPromote' ]
-      }
-
-  | otherwise
-  = GuardedSkip Nothing
-  where
-    numEstablishedPeers, numActivePeers, numPromoteInProgress :: Int
-    numEstablishedPeers  = Map.size establishedPeers
-    numActivePeers       = Set.size activePeers
-    numPromoteInProgress = Set.size inProgressPromoteWarm
-    numDemoteInProgress  = Set.size inProgressDemoteWarm
-
-
-jobPromoteWarmPeer :: forall peeraddr peerconn m.
-                      (Monad m, Ord peeraddr)
-                   => PeerSelectionActions peeraddr peerconn m
-                   -> peeraddr
-                   -> peerconn
-                   -> Job m (Completion m peeraddr peerconn)
-jobPromoteWarmPeer PeerSelectionActions{activatePeerConnection}
-                   peeraddr peerconn =
-    Job job handler
-  where
-    handler :: SomeException -> Completion m peeraddr peerconn
-    handler e =
-      --TODO: decide what happens if promotion fails, do we stay warm or go to
-      -- cold? Will this be reported asynchronously via the state monitoring?
-      Completion $ \st _now -> Decision {
-        decisionTrace = TracePromoteWarmFailed peeraddr e,
-        decisionState = st {
-                          inProgressPromoteWarm = Set.delete peeraddr
-                                                    (inProgressPromoteWarm st)
-                        },
-        decisionJobs  = []
-      }
-
-    job :: m (Completion m peeraddr peerconn)
-    job = do
-      --TODO: decide if we should do timeouts here or if we should make that
-      -- the responsibility of activatePeerConnection
-      activatePeerConnection peerconn
-      return $ Completion $ \st _now -> Decision {
-        decisionTrace = TracePromoteWarmDone peeraddr,
-        decisionState = st {
-                          activePeers           = Set.insert peeraddr
-                                                    (activePeers st),
-                          establishedStatus     = Map.insert peeraddr PeerHot
-                                                    (establishedStatus st),
-                          inProgressPromoteWarm = Set.delete peeraddr
-                                                    (inProgressPromoteWarm st)
-                        },
-        decisionJobs  = []
-      }
-
-
-activePeersAboveTarget :: forall peeraddr peerconn m.
-                          (MonadSTM m, Ord peeraddr)
-                       => PeerSelectionActions peeraddr peerconn m
-                       -> PeerSelectionPolicy peeraddr m
-                       -> PeerSelectionState peeraddr peerconn
-                       -> Guarded (STM m) (Decision m peeraddr peerconn)
-activePeersAboveTarget actions
-                       PeerSelectionPolicy {
-                         policyPickHotPeersToDemote
-                       }
-                       st@PeerSelectionState {
-                         knownPeers,
-                         establishedPeers,
-                         activePeers,
-                         inProgressDemoteHot,
-                         targets = PeerSelectionTargets {
-                                     targetNumberOfActivePeers
-                                   }
-                       }
-    -- Are we above the target for number of active peers?
-    -- Or more precisely, how many active peers could we demote?
-  | let numActivePeers, numPeersToDemote :: Int
-        numActivePeers   = Set.size activePeers
-        -- The main constraint on how many to demote is the difference in the
-        -- number we have now vs the target. We must also subtract the number
-        -- we're already demoting so we don't repeat the same work.
-        numPeersToDemote = numActivePeers
-                         - targetNumberOfActivePeers
-                         - Set.size inProgressDemoteHot
-  , numPeersToDemote > 0
-  = Guarded Nothing $ do
-
-      let availableToDemote :: Map peeraddr KnownPeerInfo
-          availableToDemote = KnownPeers.toMap knownPeers
-                               `Map.restrictKeys` activePeers
-                               `Map.withoutKeys` inProgressDemoteHot
-      selectedToDemote <- pickPeers
-                            policyPickHotPeersToDemote
-                            availableToDemote
-                            numPeersToDemote
-      let selectedToDemote' :: Map peeraddr peerconn
-          selectedToDemote' = establishedPeers
-                                `Map.restrictKeys` selectedToDemote
-
-      return Decision {
-        decisionTrace = TraceDemoteHotPeers
-                          targetNumberOfActivePeers
-                          numActivePeers
-                          selectedToDemote,
-        decisionState = st {
-                          inProgressDemoteHot = inProgressDemoteHot
-                                             <> selectedToDemote
-                        },
-        decisionJobs  = [ jobDemoteActivePeer actions peeraddr peerconn
-                        | (peeraddr, peerconn) <- Map.assocs selectedToDemote' ]
-      }
-
-  | otherwise
-  = GuardedSkip Nothing
-
-jobDemoteActivePeer :: forall peeraddr peerconn m.
-                       (Monad m, Ord peeraddr)
-                    => PeerSelectionActions peeraddr peerconn m
-                    -> peeraddr
-                    -> peerconn
-                    -> Job m (Completion m peeraddr peerconn)
-jobDemoteActivePeer PeerSelectionActions{deactivatePeerConnection}
-                    peeraddr peerconn =
-    Job job handler
-  where
-    handler :: SomeException -> Completion m peeraddr peerconn
-    handler e =
-      -- It's quite bad if closing fails, but the best we can do is revert to
-      -- the state where we believed these peers are still warm, since then we
-      -- can have another go at the ones we didn't yet try to close, or perhaps
-      -- it'll be closed for other reasons and our monitoring will notice it.
-      Completion $ \st _now -> Decision {
-        decisionTrace = TraceDemoteHotFailed peeraddr e,
-        decisionState = st {
-                          inProgressDemoteHot = Set.delete peeraddr
-                                                  (inProgressDemoteHot st)
-                        },
-        decisionJobs  = []
-      }
-
-    job :: m (Completion m peeraddr peerconn)
-    job = do
-      deactivatePeerConnection peerconn
-      return $ Completion $ \st _now -> Decision {
-        decisionTrace = TraceDemoteHotDone peeraddr,
-        decisionState = st {
-                          activePeers         = Set.delete peeraddr
-                                                  (activePeers st),
-                          establishedStatus   = Map.insert peeraddr PeerWarm
-                                                  (establishedStatus st),
-                          inProgressDemoteHot = Set.delete peeraddr
-                                                  (inProgressDemoteHot st)
-                        },
-        decisionJobs  = []
-      }
-
-
-
-changedLocalRootPeers :: forall peeraddr peerconn m.
-                         (MonadSTM m, Ord peeraddr)
-                      => PeerSelectionActions peeraddr peerconn m
-                      -> PeerSelectionState peeraddr peerconn
-                      -> Guarded (STM m) (Decision m peeraddr peerconn)
-changedLocalRootPeers actions@PeerSelectionActions{readLocalRootPeers}
-                      st@PeerSelectionState{
-                        localRootPeers,
-                        publicRootPeers,
-                        knownPeers,
-                        establishedPeers,
-                        activePeers,
-                        inProgressDemoteHot,
-                        targets = PeerSelectionTargets{targetNumberOfKnownPeers}
-                      } =
-    Guarded Nothing $ do
-      -- We have to enforce the invariant that the number of root peers is
-      -- not more than the target number of known peers. It's unlikely in
-      -- practice so it's ok to resolve it arbitrarily using Map.take.
-      localRootPeers' <- Map.take targetNumberOfKnownPeers <$> readLocalRootPeers
-      check (localRootPeers' /= localRootPeers)
-
-      let added       = localRootPeers' Map.\\ localRootPeers
-          removed     = localRootPeers  Map.\\ localRootPeers'
-          addedSet    = Map.keysSet added
-          removedSet  = Map.keysSet removed
-          knownPeers' = KnownPeers.insert PeerSourceLocalRoot
-                                          (added Map.!)
-                                          addedSet
-
-                        -- We do not immediately remove old ones from the
-                        -- known peers set because we may have established
-                        -- connections, but we mark them so that policy
-                        -- functions can prioritise them to forget:
-                      . KnownPeers.insert PeerSourceStaleRoot
-                                          (const DoNotAdvertisePeer)
-                                          removedSet
-                      $ knownPeers
-
-          -- We have to adjust the publicRootPeers to maintain the invariant
-          -- that the local and public sets are non-overlapping.
-          publicRootPeers' = publicRootPeers Set.\\ Map.keysSet localRootPeers'
-
-          -- If we are removing local roots and we have active connections to
-          -- them then things are a little more complicated. We would typically
-          -- change local roots so that we can establish new connections to
-          -- the new local roots. But since we will typically already be at our
-          -- target for active peers then that will not be possible without us
-          -- taking additional action. What we choose to do here is to demote
-          -- the peer from active to warm, which will then allow new ones to
-          -- be promoted to active.
-          selectedToDemote  :: Set peeraddr
-          selectedToDemote' :: Map peeraddr peerconn
-
-          selectedToDemote  = activePeers `Set.intersection` removedSet
-          selectedToDemote' = establishedPeers
-                               `Map.restrictKeys` selectedToDemote
-      return Decision {
-        decisionTrace = TraceLocalRootPeersChanged localRootPeers
-                                                   localRootPeers',
-        decisionState = st {
-                          localRootPeers      = localRootPeers',
-                          publicRootPeers     = publicRootPeers',
-                          knownPeers          = knownPeers',
-                          inProgressDemoteHot = inProgressDemoteHot
-                                             <> selectedToDemote
-                        },
-        decisionJobs  = [ jobDemoteActivePeer actions peeraddr peerconn
-                        | (peeraddr, peerconn) <- Map.assocs selectedToDemote' ]
-      }
-
-
-changedTargets :: MonadSTM m
-               => PeerSelectionActions peeraddr peerconn m
-               -> PeerSelectionState peeraddr peerconn
-               -> Guarded (STM m) (Decision m peeraddr peerconn)
-changedTargets PeerSelectionActions{readPeerSelectionTargets}
-               st@PeerSelectionState{
-                 localRootPeers,
-                 targets
-               } =
-    Guarded Nothing $ do
-      targets' <- readPeerSelectionTargets
-      check (targets' /= targets)
-
-      -- We have to enforce the invariant that the number of root peers is
-      -- not more than the target number of known peers. It's unlikely in
-      -- practice so it's ok to resolve it arbitrarily using Map.take.
-      let localRootPeers' = Map.take (targetNumberOfKnownPeers targets')
-                                     localRootPeers
-
-      return Decision {
-        decisionTrace = TraceTargetsChanged targets targets',
-        decisionJobs  = [],
-        decisionState = assert (sanePeerSelectionTargets targets')
-                        st {
-                          targets        = targets',
-                          localRootPeers = localRootPeers'
-                        }
-      }
-
-
-jobCompleted :: MonadSTM m
-             => JobPool m (Completion m peeraddr peerconn)
-             -> PeerSelectionState peeraddr peerconn
-             -> Time
-             -> Guarded (STM m) (Decision m peeraddr peerconn)
-jobCompleted jobPool st now =
-    -- This case is simple because the job pool returns a 'Completion' which is
-    -- just a function from the current state to a new 'Decision'.
-    Guarded Nothing $ do
-      Completion completion <- JobPool.collect jobPool
-      return $! completion st now
-
-
-monitorConnections :: forall m peeraddr peerconn.
-                      (MonadSTM m, Ord peeraddr)
-                   => PeerSelectionActions peeraddr peerconn m
-                   -> PeerSelectionState peeraddr peerconn
-                   -> Guarded (STM m) (Decision m peeraddr peerconn)
-monitorConnections PeerSelectionActions{monitorPeerConnection}
-                   st@PeerSelectionState {
-                     activePeers,
-                     establishedPeers,
-                     establishedStatus,
-                     inProgressDemoteHot,
-                     inProgressDemoteWarm
-                   } =
-    Guarded Nothing $ do
-      establishedStatus' <- traverse monitorPeerConnection establishedPeers
-      let demotions = asynchronousDemotions establishedStatus
-                                            establishedStatus'
-      check (not (Map.null demotions))
-      let (demotedToWarm, demotedToCold) = Map.partition (==PeerWarm) demotions
-      return Decision {
-        decisionTrace = TraceDemoteAsynchronous demotions,
-        decisionJobs  = [],
-        decisionState = st {
-                          activePeers       = activePeers
-                                                Set.\\ Map.keysSet demotedToWarm,
-                          establishedPeers  = establishedPeers
-                                                Map.\\ demotedToCold,
-
-                          -- Note that we do not use establishedStatus' which
-                          -- has the synchronous ones that are supposed to be
-                          -- handled elsewhere. We just update the async ones:
-                          establishedStatus = establishedStatus <> demotions
-                        }
-      }
-  where
-    -- Those demotions that occurred not as a result of action by the governor.
-    -- They're further classified into demotions to warm, and demotions to cold.
-    asynchronousDemotions :: Map peeraddr PeerStatus
-                          -> Map peeraddr PeerStatus
-                          -> Map peeraddr PeerStatus
-    asynchronousDemotions old new =
-      Map.mapMaybeWithKey asyncDemotion
-        (Map.filter (uncurry (>))
-           (Map.intersectionWith (,) old new))
-
-    -- The asynchronous ones, those not directed by the governor, are:
-    -- hot -> warm, warm -> cold and hot -> cold, other than the ones in the in
-    -- relevant progress set.
-    asyncDemotion :: peeraddr -> (PeerStatus, PeerStatus) -> Maybe PeerStatus
-    asyncDemotion peeraddr (PeerHot, PeerWarm)
-      | peeraddr `Set.notMember` inProgressDemoteHot  = Just PeerWarm
-    asyncDemotion peeraddr (PeerWarm, PeerCold)
-      | peeraddr `Set.notMember` inProgressDemoteWarm = Just PeerCold
-    asyncDemotion _        (PeerHot, PeerCold)        = Just PeerCold
-    asyncDemotion _        _                          = Nothing
 
 
 ------------------------
@@ -1919,31 +562,3 @@ peerChurnGovernor :: MonadSTM m
                   -> m () --Void
 peerChurnGovernor _ =
     return ()
-
-
--------------------------------
--- Utils
---
-
--- | Perform a first-to-finish synchronisation between:
---
--- * /all/ the async actions completing; or
--- * the timeout with whatever partial results we have at the time
---
--- The result list is the same length and order as the asyncs, so the results
--- can be paired up.
---
-waitAllCatchOrTimeout :: (MonadAsync m, MonadTimer m)
-                      => [Async m a]
-                      -> DiffTime
-                      -> m (Either [Maybe (Either SomeException a)]
-                                   [Either SomeException a])
-waitAllCatchOrTimeout as time = do
-    t       <- newTimeout time
-    results <- atomically $
-                         (Right <$> mapM waitCatchSTM as)
-                `orElse` (Left  <$> (awaitTimeout t >> mapM pollSTM as))
-    case results of
-      Right{} -> cancelTimeout t
-      _       -> return ()
-    return results
